requirements.txt
Flask==2.0.2
gunicorn==20.1.0
requests==2.11.1
pymongo[srv]
python-dotenv==0.19.2